+++ src/package/themes/default/components/container.ts
@@ -1,0 +1,10 @@
+export default {
+    maxWidth: '72rem',
+    maxWidthSm: '40rem',
+    maxWidthMd: '48rem',
+    maxWidthLg: '64rem',
+    maxWidthXl: '80rem',
+    paddingX: '1rem',
+    paddingXSm: '1.25rem',
+    paddingXLg: '1.5rem',
+};